+++ src/main/java/guess/number/GuessNumber.java
@@ -3,10 +3,10 @@ package guess.number;
 import java.util.Scanner;
 
 public class GuessNumber {
-    public static void guess(int left_border, int right_border) {
+    public static void guess(long left_border, long right_border) {
         Scanner scanner = new Scanner(System.in);
         while (right_border - left_border > 1) {
-            int middle = (right_border + left_border) / 2;
+            long middle = (right_border + left_border) / 2;
             System.out.println("Is your number less than " + middle + "?");
             String ans = scanner.next();
             if (ans.equals("yes")) {
